// Practica5/frontend/yo-usac-frontend/src/types/grabaciones.types.ts
export interface Tema {
  id_tema: number
  id_unidad: number
  nombre: string
  descripcion: string
  unidad?: string
}

export interface ClaseGrabada {
  id_clase: number
  id_curso: number
  id_periodo: number
  id_area: number
  titulo: string
  fecha_impartida: string
  duracion_min: number
  descripcion: string
  url_video: string
  anio: number
  num_semestre: number
}

export interface CatalogoClase {
  id_clase: number
  titulo: string
  descripcion: string
  fecha_impartida: string
  duracion_min: number
  url_video: string
  anio: number
  num_semestre: number
  id_curso: number
  id_area: number
  id_periodo: number
}

export interface MaterialApoyo {
  id_material: number
  id_clase: number
  nombre: string
  tipo: string
  url: string
}

export interface Participante {
  id_clase: number
  id_usuario: number
  tipo_participante: string
}

export interface DetalleClaseGrabada {
  clase: ClaseGrabada
  temas: Tema[]
  materiales: MaterialApoyo[]
  participantes: Participante[]
}
// ...
export interface Capitulo {
  id_capitulo: number
  id_clase: number
  titulo: string
  tiempo_inicio: number
}